+++ packages/ts-lib-tools/src/config/paths.ts
@@ -88,6 +88,22 @@ export const DIST_PATH = resolveRoot('./dist');
 export const MODULE_TEMPLATE_PATH = resolve(ASSETS_PATH, './module-template');
 
 /**
+ * 获取初始化测试文件
+ */
+export const getTestSetups = () => resolveModule(resolveRoot, 'src/setupTests');
+
+/**
+ * 获取 @testing-library/jest-dom 模块的路径
+ */
+export const getJestDOMModulePath = () => {
+  try {
+    return require.resolve('@testing-library/jest-dom/extend-expect');
+  } catch {
+    return undefined;
+  }
+};
+
+/**
  * 库对应的全局名称
  */
 export const globals = () => {