+++ packages/ts-lib-jsdom-polyfill/index.js
@@ -1,0 +1,8 @@
+if (typeof window !== 'undefined') {
+  window.requestAnimationFrame = (callback) => {
+    return setTimeout(callback, 16);
+  };
+  window.cancelAnimationFrame = (rafId) => {
+    clearTimeout(rafId);
+  };
+}